refactor: cleanup outdated tax pages and improve export robustness
- Remove TaxPageNew and GovSubmissionsPage (replaced by ComprehensiveTaxPage)
- Remove /tax and /gov-submissions routes from main.dart
- Add defensive Number() casts in ExportService.generateKRA_P10_CSV
- Streamline navigation: Finance tab -> Tax Filing -> ComprehensiveTaxPage
- Clean up debug SQL files and test PowerShell scripts
- Update deploy-prod workflow

+++ backend/src/modules/workers/entities/worker-document.entity.ts
@@ -43,7 +43,7 @@ export class WorkerDocument {
     @Column()
     url: string; // File URL
 
-    @Column({ type: 'bigint', nullable: true })
+    @Column({ type: 'integer', nullable: true })
     fileSize: number; // Size in bytes
 
     @Column({ nullable: true })
